Adding CollapseSceneElement and TransformableSceneElement:
CollapseSceneElement: Decorates a SceneElement so that it is treated as a single primitive.  This allows the SceneElement to be inserted into a BVH scene element (e.g., BoundingIntervalHierarchy) as a single entity.
TransformableSceneElement:  Decorates a SceneElement allowing it to be transformed according to an affine transformation.  This is needed for instancing.