Add MatrixBuffer to store Matrix elements
Rather than having Matrix directly maintain the backing array and
offsets/strides, we extract this responsibility to a MatrixBuffer class,
which effectively acts as a mutable Matrix.  In this way, we now have a
class that supports the operations of a Matrix but allow for in-place
computations, while maintaining immutability of the Matrix class (at
least, when the instance's underlying MatrixBuffer is not available).