Add ProtoRenderJob to pass through job as serialized proto.
ProtoRenderJob maintains a reference to the RenderJob proto used to
create it, expands the proto into a transient ParallelizableJob on
demand.  This way, only the proto is sent over the wire and the
implementation is recreated on the other side.  As a result, the classes
comprising the implementation do not need to be Serializable, and we
save the unnecessary step of expanding and reserializing a job that was
provided in a serializable form (i.e., a protobuf) to begin with.