Traverse children when ray aligned with planes.
If the ray is exactly perpendicular to the planes for the current level
of a BIH, we end up dividing by zero when trying to determine the
distances to the ray-intersections with the planes. We should avoid this
and instead just traverse both children if the ray is too close to
alignment with the planes.