Use ThreadLocalRandom for default random generator
Math.random() performs poorly when used heavily in highly-multithreaded
environments. It essentially works by attempting, in a loop, to generate
a new random number and then update the seed -- under the (optimistic)
assumption that no other threads are trying to do the same concurrently.
The loop repeats until this assumption is satisfied, i.e., until it can
update the seed without finding that another thread has already done so.

ThreadLocalRandom does not have this problem because a separate seed is
used for each thread.